refactor: lotto 패키지로 도메인, 뷰, 컨트롤러 이동
