+++ src/main/kotlin/Application.kt
@@ -1,6 +1,6 @@
-import controller.LottoController
-import view.InputView
-import view.OutputView
+import lotto.controller.LottoController
+import lotto.view.InputView
+import lotto.view.OutputView
 
 fun main() {
     val inputView = InputView()
