+++ src/main/kotlin/lotto/model/LottoNumber.kt
@@ -1,4 +1,4 @@
-package domain.model
+package lotto.model
 
 @JvmInline
 value class LottoNumber(val value: Int) {
