+++ src/main/kotlin/lotto/model/Margin.kt
@@ -1,3 +1,3 @@
-package domain.model
+package lotto.model
 
 data class Margin(val rate: Double)
