+++ src/test/kotlin/domain/LottoTest.kt
@@ -1,5 +1,7 @@
 package domain
 
+import lotto.model.Lotto
+import lotto.model.MakeLottoStrategy
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
