+++ src/test/kotlin/domain/RankTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import lotto.model.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
