+++ src/test/kotlin/domain/model/LottoNumberTest.kt
@@ -1,5 +1,6 @@
 package domain.model
 
+import lotto.model.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
