+++ src/test/kotlin/domain/model/MoneyTest.kt
@@ -1,5 +1,6 @@
 package domain.model
 
+import lotto.model.Money
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
